[*] Mudança de double para decimal nos campos do Sintegra, mantendo o padrão com os demais componentes do ACBrFramework.Net - Rafael Batiati

+++ ACBrFramework/ACBrFramework.Net/ACBrFramework.Net/Sintegra/Registro/Registro 51/SintegraRegistro51.cs
@@ -18,14 +18,14 @@ namespace ACBrFramework.Sintegra
 
 		public DateTime DataDocumento { get; set; }
 
-		public double ValorIpi { get; set; }
+		public decimal ValorIpi { get; set; }
 
-		public double ValorContabil { get; set; }
+		public decimal ValorContabil { get; set; }
 
 		public string Situacao { get; set; }
 
-		public double ValorIsentas { get; set; }
+		public decimal ValorIsentas { get; set; }
 
-		public double ValorOutras { get; set; }
+		public decimal ValorOutras { get; set; }
 	}
 }
